+++ lib/language_pack/ruby_version.rb
@@ -38,6 +38,13 @@ module LanguagePack
       @version_without_patchlevel = @version.sub(/-p-?\d+/, '')
     end
 
+    def warn_ruby_26_bundler?
+      return false if Gem::Version.new(self.ruby_version) >= Gem::Version.new("2.6.3")
+      return false if Gem::Version.new(self.ruby_version) < Gem::Version.new("2.6.0")
+
+      return true
+    end
+
     def ruby_192_or_lower?
       Gem::Version.new(self.ruby_version) <= Gem::Version.new("1.9.2")
     end